Reorganize test modules to prevent parallel execution during pants run
